requirements.txt
google-cloud-aiplatform[agent_engines,adk]>=1.112
google-adk>=1.0.0
google-auth>=2.0.0
requests>=2.31.0
fredapi>=0.5.0
praw>=7.7.0
